+++ blocks/accordion/accordion.css
@@ -1,17 +1,17 @@
 .accordion details {
-  border-bottom: 1px solid #C4C4C4;
+  border-bottom: 0.063rem solid var(--color-neutral-300);
 }
 
 /* stylelint-disable-next-line no-descending-specificity */
 .accordion details + details {
-  margin-top: 24px;
+  margin-top: 1.5rem;
 }
 
 .accordion details p {
   margin-bottom: 0.8em;
-  padding-left: 20px;
-  font-family: Gilmer, sans-serif;
-  font-size: 16px;
+  padding-left: 1.25rem;
+  font-family: var(--font-regular);
+  font-size: 1rem;
   font-style: normal;
   line-height: 150%;
 }
@@ -33,12 +33,12 @@
   content: '';
   position: absolute;
   top: 50%;
-  left: 2px;
+  left: 0.125rem;
   transform: translateY(-50%) rotate(135deg);
-  width: 6px;
-  height: 6px;
-  border: 2px solid;
-  border-width: 2px 2px 0 0;
+  width: 0.375rem;
+  height: 0.375rem;
+  border: 0.125rem solid;
+  border-width: 0.125rem 0.125rem 0 0;
   transition: transform 0.2s;
 }
 
@@ -52,7 +52,7 @@
 }
 
 .accordion details .accordion-item-body {
-  padding: 0 16px;
+  padding: 0 1rem;
   padding: 0;
 }
